refactor: split sync.go and config.go along their topic seams
Pure declaration moves, no code changes — every block travels with its
doc comment; imports re-derived per file.

sync.go (2628 lines) → sync.go (orchestration + phases, 575),
sync_discover.go (discovery + drop/research collection),
sync_extract.go (project extraction), sync_sessions.go (ccrider
mining), sync_absorb.go (raw-article absorb).

config.go (2070) → config.go (ScribeConfig + loading + misc blocks,
765), config_llm.go (per-op LLM routing + inheritance helpers),
config_absorb.go (absorb/contextualize/facts knobs), frontmatter.go
(frontmatter parsing + markdown walkers — never config to begin with).
